+++ changelog.md
@@ -1,3 +1,7 @@
+# Version 0.4.1
+- Added error handling for window settings
+- Added error handling for connection file
+
 # Version0.4
 - Editing connections now possible
 - Delete connections now possible